backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
google-genai>=1.0.0
supabase>=2.0.0
httpx>=0.28.0
pydantic>=2.0.0
elevenlabs>=2.0.0
PyJWT>=2.8.0
